重症患者数: Rename components/cards/SevereCaseCard.vue to components/index/CardsMonitoring/SevereCase/Card.vue
