+++ components/index/CardsMonitoring/SevereCase/Card.vue
